requirements.txt
Flask==1.1.1
waitress==1.3.0
python-decouple==3.1
